feat: Traderbot V2
Impementado traderbot v2 com base no arquivo ModeloBTC01.py
