+++ .idea/runConfigurations/traderbot_v1___train_data.xml
@@ -23,7 +23,7 @@
         <ENTRY IS_ENABLED="true" PARSER="runconfig" IS_EXECUTABLE="false" />
       </ENTRIES>
     </EXTENSION>
-    <option name="SCRIPT_NAME" value="$PROJECT_DIR$/src/traderbot_v1/treinamento_ia.py" />
+    <option name="SCRIPT_NAME" value="$PROJECT_DIR$/src/traderbot_v1/ai_training.py" />
     <option name="PARAMETERS" value="" />
     <option name="SHOW_COMMAND_LINE" value="false" />
     <option name="EMULATE_TERMINAL" value="false" />
