+++ src/traderbot_v2/ai_training.py
